Clears cached tokens in TokenHolder before loading new ones
instead of during stop since there's a window of opportunity for a
concurrent shutdown and a transaction racing so that the thread shutting
down makes it as far as clearing the token maps right before the
transaction thread tries to create a relationship of a certain type, sees
that it's not there and decides to create it. At this point the
transaction thread got past the availability guard right before it shut
down and at the point where it's at for the moment there are enough
components alive in the database to have the transaction complete.

The end result would be that there would now exist duplicate tokens with
the same name, but different ids, something that would be observed the
next startup.

This scenario is very hard to test for with making a brittle and
potentially flaky test.