Refactor the counts store
This gives the count store a self-describing format header, that encodes
the key size and value size, as well as a descriptor that can be used to
know what the serialisation format of the keys and values are. This will
allow us to upgrade the store format of the counts store seamlessly, and
distinguish between the case of a count not existing because it is zero,
and a count not existing because the store is of a version that does not
support that particular count.

Along with the change to the header format, a trailer was added to the
store, so that the counts store now has a versioned trailer just like
all the other store files, and will not need any special treatment
during store migration.

This commit also includes a change to make application of transactions
to the count store atomic, by introducing an "Updater" object for the
store that will hold a lock preventing rotation while a change is being
applied. This prevents issues in store-copying where the copied state of
the count store could include partial transactions.