Importer page cache uses larger page size for relationship store
There was a problem, at least on Windows 7, where the OS would somehow
keep the entire file, or very large portions of it in memory when reading
the relationship store backwards as done in the final stage of the import.
This would be a performance problem where releasing of this memory would
come first when the import was done, or all available RAM was used at
which point the OS would start releasing some parts of the cached
relationship file. Although at this point the import process would observe
significant stalls and slowdowns.

The solution I've seen is to, when reading a file backwards, read it in
much bigger chunks. When doing so the OS doesn't seem to cache the file
like described above.

A pretty good description of the problem:
http://stackoverflow.com/questions/18171981/reading-file-with-fread-in-reverse-order-causes-memory-leak/