Lets CharReadable/CharSeeker able to provide source information
such as name of file, if such would be the current data source, as well as
position and line number. Meant for human consumption, for inclusion in
error messages and similar.

This information is used by the import tool to display information on in
which file and line number f.ex. duplicate node ids are defined.