Releases KernelTransactionImplementation last thing in close()
Previously a KTI was released back to its pool between completing
commit/rollback which happens in close(), and fully completing close().
As it so happens the rest of the close() was recently changed to use some
of the state inside the transaction and so releasing too soon enabled a
race where the transaction object might have gotten reused before fully
closed.