Makes FlippableIndexProxy delegate volatile
to fix a race condition where a constraint index was flipped to f.ex
an OnlineIndexProxy and right after that another thread comes in and f.ex
tries to set a node property. Then it may see it in SchemaCache, otherwise
load from disk and then assert that the state is ONLINE. At this point the
thread setting the node property might still see the old reference to the
populating index proxy, so the state will be POPULATING and the lookup
will fail.