Make StoreCopyClient use existing page cache
StoreCopyClient is used when secondaries in a cluster copies the entire store from the primary.
This works (in the short version of the story) by first marking the current transaction log position on the primary.
Then copying over all the store files, and then copying over all the transaction log contents from the mark from before.
A temporary database is then started on the store to perform recovery upon it, using the log we created.
This temporary database used to create its own page cache, possibly using the default size, or the same size as is configured for the already running database.
This is a terribly wasteful use of memory, since the default setting will attempt to use the majority of the memory available to the machine.
With these changes, the temporary database now borrows the page cache of the host database, thus wasting no excess memory, as it is now confined to not only the same heap, but also the same direct-memory space.