Check that canned cypher statements can be executed.
The browser code includes a number of canned example cypher statements.
It's important that these statements continue to work as the database
and the examples continue to evolve. This test provides the most basic
coverage: it asserts that the statements can be executed without
throwing an exception. There is no attempt to assert that the statements
have an expected behaviour or return correct results. However, this test
should be good enough to catch the majority of accidental syntax errors
in the browser code. It may also provide early warning that examples
need to be updated when cypher syntax moves on.