Monitors store copy on server and client individually
so that
BackupServiceIT#shouldContainTransactionsThatHappenDuringBackupProcess can
be implemented correctly and test what it was initially designed to test,
namely that a store copy gets correct even in the face of changing store
files while copying.

Also fixed a transaction log header issue where the txId would be 0
instead of 1 (base txId) if no transaction streamed.

thanks to @davidegrohmann