Avoids IndexingService monitor on commit if no index changes
Committing a transaction goes through a number of appliers where one of
them, the IndexTransactionApplier synchronizes on IndexingService for
every commit. This is unnecessary if no index changes are made in the
transaction. So this commit adds a ValidatedIndexUpdate#hasChanges() where
the monitor is acquired, followed by #flush(), if that returns true. This
should reduce contention when committing otherwise uncontented data in
parallel.

Also removed an eager instantiation of an ArrayList in
IndexTransactionApplier.