Fixes a race between committing transaction and db shutting down
There was a window of opportunity in shutdown (NeoStoreDataSource#stop()
specifically) after waiting for committing transactions to be closed
and before he log version was incremented where a transaction that had
gotten past the availability guard in beginTx() could commit and apply
its changes to store and return OK to the user without the transaction
being covered by the transaction log, since the log would already
have been rotated away, by the incremented log version. A crash after this
point would not recover this transaction. Actually, such a transaction
could race with the shutdown thread all the way down to unmapping the
stores from the PageCache.

This commit introduces a monitor around awaiting transactions to commit,
forcing and incrementing the log version so that this race cannot happen.

co-author: @thobe