Exposes means of correctly marking transactions as committed
TransactionAppender#append(TransactionRepresentation,long) marked appended
transactions as committed before exiting the method. This was a problem
when batching transactions, committing/applying multiple at once because
transactions would be marked as committed and potentially never reaching
the point where they were marked as closed. This would be a problem in
shutdown/rotation where committed transactions are awaited to be closed
indefinitely.

Now that append method returns a Commitment which the caller is
responsible to use, for marking that transaction as committed, after that
transaction has been forced. This makes it possible to have the marking of
committed and closed within the same try-finally block (committed in try,
closed in finally) which is a must.

TransactionCommittingResponseUnpacker now takes advantage of this properly.