Skeleton of Socket transport
Does not yet support the actual message protocol, this is the module skeleton
and a proposal for version negotiation. Choosing a version needs to be highly
flexible for future protocol changes. This lets the client propose up to four
supported protocol versions, in order of preference, and the server then chooses
the first one it supports (or tells the client none will work).

This is done outside the scope of PackStream (allowing it to change cross
versions), the messaging protocol and the (yet-to-be-created) framing protocol.

The number of proposed versions is bounded to 4 (if the client wants to try
more it can simply send a new version proposal, having it be fixed-width lowers
complexity), each version is a 4-byte unsigned integer, bounding the number of
protocol versions we can have to 4 294 967 295, which is a very high number and
should be ample time for Elon Musk to sort all this 'computer' stuff out for us.

See SocketSessionIT for examples of negotiating protocol version.