Use flat-combining for concurrent application of updates to the LabelScanStore
This is similar to how we amortise the transaction log fsync cost.
When a transaction processing thread wants to write changes to the LabelScanStore (this is during commit, after they have written their commands to the transaction log), it starts out by adding a unit of work to a concurrent, wait-free batching stack.
Then it attempts to grab a lock (this happens inside the new WorkSync abstraction).
If it gets the lock, then it grabs the entire stack of work in one go, and combines all the units of work into a single unit.
This single unit of work is then applied to a single LabelScanWriter.
This amortises the creation and closure of LabelScanWriters, which are both very costly operations compared to just writing to a LabelScanWriter.