Reduced memory usage in LuceneLegacyIndex when querying
Previously stored all returned ids in set causing out of memory for
big queries. Now only stores ids for documents in transaction state << all
ids. When an id is in transaction state, this is the most up to date
version and this is what we want to return to the user.
We previously stored all returned ids in set to avoid returning same id multiple
times. This can only accure when id is in transaction state and thus we
only need to store those ids.

co-author: @burqen