Fixes a recovery issue with legacy indexes
Where updates could go to the wrong index. This was introduced by a recent
change where commits were batched during recovery. The core of the problem
was that index name ids are transaction-local, but was treated as global,
which became a problem for appliers that used the index name id as key to
redirect updates to the right indexes.

Also previously there were one applier created for every index, even
though legacy index appliers are designed to handle multiple indexes each.
This has been fixed.