Aborts attempt to join cluster on recognizing foreign store
Attempting to join a cluster and seing that the master
(representing rest of the cluster) has a different store ID
and where the joining member has a non-empty database now throws a
specific exception which gets recognized as so serious that further join
attempts are aborted.

For this introduced a HighAvailabilityStoreFailureException, which gathers
all StoreFailureException subclasses used in HA only, under it.
SwitchToSlave now treats all such HighAvailabilityStoreFailureException as
serious errors resulting in aborting further attempts to join that cluster
without manual intervention.