Introduce code generation library
This code generation library builds on an abstraction close to Java (the language),
with methods representing the statements of Java that accept (a tree of) objects
representing the expressions of Java.

Currently there is one backend implemented for this code generation library, one
that generates java source code and compiles it through javac.